Rename usercode_display_widget->device_flow_widget
