+++ packages/google_sign_in/lib/google_sign_in_tizen.dart
@@ -6,20 +6,20 @@ import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 import 'package:google_sign_in_platform_interface/google_sign_in_platform_interface.dart';
 
+import 'src/device_flow_widget.dart' as device_flow_widget;
 import 'src/google_sign_in.dart';
-import 'src/usercode_display_widget.dart' as device_flow;
 
 /// Sets a custom [GlobalKey\<NavigatorState>] object used for pushing a Flutter
 /// widget that displays "user_code" and "verification_uri".
 void setNavigatorKey(GlobalKey<NavigatorState> key) =>
-    device_flow.navigatorKey = key;
+    device_flow_widget.navigatorKey = key;
 
 /// Returns a [GlobalKey\<NavigatorState>] object currently used for pushing a
 /// Flutter widget that displays "user_code" and "verification_uri".
 ///
 /// A default key used by the plugin will be returned unless a custom key was
 /// provided with [setNavigatorKey].
-GlobalKey<NavigatorState> getNavigatorKey() => device_flow.navigatorKey;
+GlobalKey<NavigatorState> getNavigatorKey() => device_flow_widget.navigatorKey;
 
 /// Sets [clientId] and [clientSecret] to be used for GoogleSignIn authentication.
 ///
@@ -60,7 +60,7 @@ class GoogleSignInTizen extends GoogleSignInPlatform {
             "in google_sign_in_tizen.dart before calling GoogleSignIn's signIn API.",
       );
     }
-    if (device_flow.navigatorKey.currentContext == null) {
+    if (device_flow_widget.navigatorKey.currentContext == null) {
       throw PlatformException(
         code: 'navigatorkey-unassigned',
         message: 'Cannot initialize GoogleSignInTizen: a default or custom '
