+++ packages/google_sign_in/lib/src/device_flow_widget.dart
